# replace globby with much faster code when gitignore enabled