--- a/packages/knip/src/PrincipalFactory.ts
+++ b/packages/knip/src/PrincipalFactory.ts
@@ -2,7 +2,6 @@ import ts from 'typescript';
 import { ProjectPrincipal } from './ProjectPrincipal.js';
 import { toAbsolute } from './util/path.js';
 import type { SyncCompilers, AsyncCompilers } from './types/compilers.js';
-import type { GlobbyFilterFunction } from 'globby';
 
 type Paths = ts.CompilerOptions['paths'];
 
@@ -15,7 +14,7 @@ export type PrincipalOptions = {
   paths: Paths;
   compilers: [SyncCompilers, AsyncCompilers];
   pkgName: string;
-  isGitIgnored: GlobbyFilterFunction;
+  isGitIgnored: (path: string) => boolean;
   isIsolateWorkspaces: boolean;
 };