--- a/packages/knip/src/ProjectPrincipal.ts
+++ b/packages/knip/src/ProjectPrincipal.ts
@@ -16,7 +16,6 @@ import type { SyncCompilers, AsyncCompilers } from './types/compilers.js';
 import type { ExportItem, ExportItemMember } from './types/exports.js';
 import type { UnresolvedImport } from './types/imports.js';
 import type { BoundSourceFile, GetResolvedModule, ProgramMaybe53 } from './typescript/SourceFile.js';
-import type { GlobbyFilterFunction } from 'globby';
 
 // These compiler options override local options
 const baseCompilerOptions = {
@@ -57,7 +56,7 @@ export class ProjectPrincipal {
   // We don't want to report unused exports of config/plugin entry files
   skipExportsAnalysis = new Set<string>();
 
-  isGitIgnored: GlobbyFilterFunction;
+  isGitIgnored: (path: string) => boolean;
   cwd: string;
   compilerOptions: ts.CompilerOptions;
   extensions: Set<string>;