The current dir walking method has two issues:

1. gitignores are searched for and parsed for every single plugin for every single workspace package. This search is very inefficient and walks the full directory tree, without any early exiting / pruning.
2. the gitignores are applied in memory only after walking (again, no early exiting / pruning).

The reason for both is in the globby() calls.

On every globby(patterns) call, previously this happened:

1. The full repo was recursively walked with a fast-glob glob of '**/.gitignore' to find the ignore files
3. All the ignore files are parsed
4. Now the fast-glob call with the patterns given in the config is executed (not the ones in gitignore!)
5. Now afterwards, all the ignores from gitignore are filtered out in memory (in a bit buggy way)

globby is just a thin wrapper around fast-glob, and the only globby feature knip uses is the gitignore function, which has caused multiple issues in the past (That's the reason the docs recommend trying to turn off gitignore detection for performance afaik?).

With this PR, instead this happens on a globby() call:

1. The repo is walked using @nodelib/fs.walk, incrementally building the gitignore list. Files in gitignored directories are not walked (that's what fast-glob uses internally, but fast-glob doesn't support dynamic ignores (https://github.com/mrmlnc/fast-glob/issues/265)).
2. The gitignore list is cached per passed cwd so they are only ever computed and parsed once.
     
     Since I don't think there's a lib that actually supports collecting multiple gitignore I wrote a small gitignore->micromatch converter. This converter is probably not perfect, but the globby solution is also wrong (I can give details if needed).
4. fast-glob is called with the patterns and ignore list from knip itself combined with the ignore list from the gitignore files. This causes fast-glob to also do early pruning which makes walking here much faster as well.



This should fix #425 and inverse the recommendation to set `gitignore: false` . Instead, knip should now be faster when ignoring more files which makes sense.

|(Rough) Benchmark | Before | After |
|---|---|---|
|My Repo with a huge gitignored directory inside | 30+minutes (didn't wait for it to finish). > 200 million syscalls | 2:02min |
|My Repo with just build artifacts | 3:20min | 1:50min|


The reason I'm submitting this here and not to globby itself is that (a) globby is written in JS not TS and (b) has multiple other functions that would probably be affected but knip doesn't care about.