@webpro 2024-01-04
Hi @phiresky, this is fantastic! 🔥 Thank you so much for doing this.

This is such a huge win, happy to replace globby and merge your PR. I don't have an issue with maintaining that here. Just a few points:

- This pull request reveals a bug in the 11ty plugin I've fixed in that plugin in the `main` branch, another win! Any chance you could rebase on top of that?
- Not sure why `npx knip` behaves like it does on Windows (https://github.com/webpro/knip/actions/runs/7400909611/job/20151486019), I will grab an old Windows machine this afternoon and look into it.
- I looked for a related issue and only now I see your comment in https://github.com/webpro/knip/issues/172 (sorry!), but it's great if this PR fixes `EACCES` errors too.
- The rest is stylistic I can tweak any time later on.
- And it will indeed be great to remove that "performance boost" section from the docs!