@webpro 2024-01-04
> I've rebased this on master.

Thanks!

> > Not sure why npx knip behaves like it does on Windows
> 
> If that's only happening since this branch then it's probably related to some path joining doing something with \ and not /. Globby I think has some logic to convert \ to / so there's probably some inconsistency now. `knip --debug` should reveal it but it's annoying to test without Windows.

I have a Windows laptop right here, I'll look into it.

> > great if this PR fixes EACCES errors too.
> 
> It should fix access errors for files / directories that are in ignored directories yes. If there's an access error in a file that's part of the repo then no. (But that seems like it probably shouldn't be ignored anyways)

Yeah that's what I meant, at least users can ignore problematic dirs/files.

> > to remove that "performance boost" section
> 
> Note that this _potentially can_ still decrease the performance against `gitignore: false` if the repo is fully clean anyways. Since (a) it still does an additional directory traversal to find gitignore files and (b) the number of patterns passed to fast-glob is larger especially if the gitignore files are large. The clean repo is something that rarely happens locally, but in a CI environment it's well possible.
> 
> I don't think the impact should be much though because it's a single scan that (on my larger repo) only takes <1s. fast-glob optimizes patterns I think by e.g. parsing them into path segments and matching those faster but not sure.

Thanks, I'll benchmark a bit. But the overall win is a no-brainer 🙏

On a related note, have you seen https://github.com/fabiospampinato/fast-ignore? If not, might have some good ideas as well.