@webpro 2024-01-04
Easy fix up for Windows, to be continued.. Just for the record:

```diff
diff --git a/packages/knip/src/util/globby.ts b/packages/knip/src/util/globby.ts
index 1fc3ef1e..2b9fb378 100644
--- a/packages/knip/src/util/globby.ts
+++ b/packages/knip/src/util/globby.ts
@@ -37,7 +37,7 @@ function convertGitignoreToMicromatch(pattern: string, base: string) {
 
 function parseGitignoreFile(filePath: string, cwd: string) {
   const file = fs.readFileSync(filePath, 'utf8');
-  const base = path.relative(cwd, path.dirname(filePath));
+  const base = path.relative(cwd, path.dirname(path.toPosix(filePath)));
 
   return file
     .split(/\r?\n/)

```